+++ PE-Editor/assets/shaders/Renderer3D_DefaultLit.vert
@@ -36,7 +36,7 @@ struct VertexData
 };
 
 layout(location = 0) out flat int v_EntityID;
-layout(location = 1) out flat int v_MaterialIndex;
+layout(location = 1) out flat uint v_MaterialIndex;
 layout(location = 2) out VertexData v_VertexData;
 
 void main()
@@ -56,7 +56,7 @@ void main()
 	v_VertexData.TBN = mat3(T, B, N);
 
 	v_EntityID = MeshSubmissions[gl_DrawID].EntityID;
-	v_MaterialIndex = MeshSubmissions[gl_DrawID].MaterialIndex;
+	v_MaterialIndex = uint(MeshSubmissions[gl_DrawID].MaterialIndex);
 
 	gl_Position = u_CameraBuffer.Projection * u_CameraBuffer.View * vec4(v_VertexData.WorldFragPos, 1.0);
 }
